requirements.txt
trello
gkeepapi
python-dotenv
azure-functions